+++ src/callbacks/__init__.py
@@ -1,0 +1,1 @@
+from .update_charts import *
