+++ src/components/__init__.py
@@ -1,0 +1,4 @@
+from .sidebar import create_sidebar
+from .footer import create_footer
+from .title import title
+from .charts import salary_trend, salary_by_experience_level, salary_by_company_size, salary_by_remote_type
