+++ src/components/title.py
@@ -1,0 +1,16 @@
+from dash import html
+
+title = html.H1(
+    '💰DataSalaries - Salary Insights for Data Professionals',
+    style={
+        'backgroundColor': 'steelblue',
+        'padding': 20,
+        'color': 'white',
+        'margin-top': 20,
+        'margin-bottom': 20,
+        'text-align': 'center',
+        'font-size': '48px',
+        'border-radius': 3,
+        'margin-left': -10,
+    }
+)
